Extract base model functionality from Article
New models should inherit from the base class Model, which provides
an abstraction over mongodb's documents. Article only implements Model,
which means defining a schema and a database name, leaving the logic to
the base class entirely.
